Simplify types and cleanup
